added styles to main page
