+++ src/App.jsx
@@ -3,8 +3,9 @@ import { useDispatch } from 'react-redux';
 import { Route, Switch } from 'react-router-dom';
 import { fetchStories, fetchTop100StoriesId } from './api/hn-api';
 import { actions as storiesActions } from './store/storiesSlice.js';
-import Main from './components/Layout/Main.jsx';
-import Header from './components/Layout/Header.jsx';
+import Wrapper from './components/Layout/Wrapper/Wrapper';
+import Main from './components/Layout/Main/Main.jsx';
+import Header from './components/Layout/Header/Header.jsx';
 import Footer from './components/Layout/Footer.jsx';
 import Home from './components/Home.jsx';
 import Story from './components/Story/Story.jsx';
@@ -23,18 +24,20 @@ const App = () => {
 
   return (
     <>
-      <Header />
-      <Main>
-        <Switch>
-          <Route exact path='/'>
-            <Home />
-          </Route>
-          <Route path='/story/:id'>
-            <Story />
-          </Route>
-        </Switch>
-      </Main>
-      <Footer />
+      <Wrapper>
+        <Header />
+        <Main>
+          <Switch>
+            <Route exact path='/'>
+              <Home />
+            </Route>
+            <Route path='/story/:id'>
+              <Story />
+            </Route>
+          </Switch>
+        </Main>
+        <Footer />
+      </Wrapper>
     </>
   );
 }
