+++ src/components/Home.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import NewsList from "./NewsList.jsx";
+import NewsList from "./NewsList/NewsList.jsx";
 
 const Home = () => {
   return (
