--- src/components/Layout/Header.jsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-const Header = () => (
-  <header>
-    <h1>Hacker News Top 100 Reader</h1>
-  </header>
-);
-
-export default Header;
