+++ src/components/Layout/Header/Header.module.css
@@ -1,10 +1,14 @@
+@import url('../../../styles/constants.css');
+
 .header {
   display: flex;
   align-items: center;
-  justify-content: space-between;
+  justify-content: center;
   margin-bottom: 2rem;
 }
 
-.header-title {
+.title {
+  font-family: Montserrat, sans-serif;
+  font-weight: 900;
   color: var(--textColor);
 }
