--- src/components/Layout/Main.jsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-const Main = (props) => <main>{props.children}</main>;
-
-export default Main;
