+++ src/components/Layout/Main/Main.jsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import styles from './Main.module.css';
+
+const Main = (props) => <main className={styles.main}>{props.children}</main>;
+
+export default Main;
