+++ src/components/Layout/Main/Main.module.css
@@ -1,0 +1,9 @@
+@import url('/src/styles/constants.css');
+
+.main {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  flex-wrap: wrap;
+  color: (--textColor);
+}
