+++ src/components/Layout/Wrapper/Wrapper.module.css
@@ -1,5 +1,5 @@
 .wrapper {
   min-height: 100vh;
   margin: 0 auto;
-  width: 1220px;
+  width: 780px;
 }
