+++ src/index.css
@@ -1,8 +1,51 @@
+@import-normalize;
+@import url('./styles/constants.css');
+
+@font-face {
+  font-family: 'Montserrat';
+  font-style: normal;
+  font-weight: 400;
+  src: local(''),
+    url('./styles/fonts/montserrat-v25-latin-regular.woff2') format('woff2'),
+    url('./styles/fonts/montserrat-v25-latin-regular.woff') format('woff');
+}
+
+@font-face {
+  font-family: 'Montserrat';
+  font-style: normal;
+  font-weight: 900;
+  src: local(''),
+    url('./styles/fonts/montserrat-v25-latin-900.woff2') format('woff2'),
+    url('./styles/fonts/montserrat-v25-latin-900.woff') format('woff');
+}
+
+@font-face {
+  font-family: 'Merriweather';
+  font-style: normal;
+  font-weight: 400;
+  src: local(''),
+    url('./styles/fonts/merriweather-v30-latin-regular.woff2') format('woff2'),
+    url('./styles/fonts/merriweather-v30-latin-regular.woff') format('woff');
+}
+
+@font-face {
+  font-family: 'Merriweather';
+  font-style: normal;
+  font-weight: 900;
+  src: local(''),
+    url('./styles/fonts/merriweather-v30-latin-900.woff2') format('woff2'),
+    url('./styles/fonts/merriweather-v30-latin-900.woff') format('woff');
+}
+
+.page {
+  font-family: 'Merriweather';
+  font-style: normal;
+  font-weight: 400;
+  background-color: var(--bg-color);
+}
+
 body {
   margin: 0;
-  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
-    'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
-    sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
 }
