+++ src/styles/constants.css
@@ -1,0 +1,5 @@
+:root {
+  --bg-color: #282c35;
+  --textColor: #fff;
+  --pink: #ffa7c4;
+}
